separated project structure in to TAF model. ready to start to work in POM
