+++ package.json
@@ -16,16 +16,16 @@
   "scripts": {
     "test:all": "npm run test:playwright",
 
-    "test:playwright": "npx playwright test",
-    "test:playwright:headed": "npx playwright test --headed",
-    "test:playwright:debug": "npx playwright test --debug",
-    "test:playwright:ui": "npx playwright test --ui",
+    "test:playwright": "npx playwright test --config=src/configs/playwright.config.ts",
+    "test:playwright:headed": "npx playwright test --config=src/configs/playwright.config.ts --headed",
+    "test:playwright:debug": "npx playwright test --config=src/configs/playwright.config.ts --debug",
+    "test:playwright:ui": "npx playwright test --config=src/configs/playwright.config.ts --ui",
 
-    "test:login": "npx playwright test tests/1-login.spec.ts",
-    "test:profile": "npx playwright test tests/2-profile.spec.ts",
-    "test:products": "npx playwright test tests/3-products-and-cart.spec.ts",
-    "test:search": "npx playwright test tests/4-search-and-filter.spec.ts",
-    "test:checkout": "npx playwright test tests/5-checkout.spec.ts",
+    "test:login": "npx playwright test --config=src/configs/playwright.config.ts src/tests/1-login.spec.ts",
+    "test:profile": "npx playwright test --config=src/configs/playwright.config.ts src/tests/2-profile.spec.ts",
+    "test:products": "npx playwright test --config=src/configs/playwright.config.ts src/tests/3-products-and-cart.spec.ts",
+    "test:search": "npx playwright test --config=src/configs/playwright.config.ts src/tests/4-search-and-filter.spec.ts",
+    "test:checkout": "npx playwright test --config=src/configs/playwright.config.ts src/tests/5-checkout.spec.ts",
     
     "show-report": "npx playwright show-report",
     "install-browsers": "npx playwright install",
@@ -42,11 +42,11 @@
   },
   "homepage": "https://github.com/ITSECT3R/AT-Specializtion-Epam#readme",
   "devDependencies": {
-    "@playwright/test": "^1.54.2",
-    "@types/chai": "^5.2.2",
-    "@types/node": "^24.3.0",
-    "chai": "^6.0.1",
-    "ts-node": "^10.9.2",
-    "tsconfig-paths": "^4.2.0"
+    "@playwright/test": "1.54.2",
+    "@types/chai": "5.2.2",
+    "@types/node": "24.3.0",
+    "chai": "6.0.1",
+    "ts-node": "10.9.2",
+    "tsconfig-paths": "4.2.0"
   }
 }
