+++ src/configs/tsconfig.json
