+++ src/po/base.page.ts
@@ -1,0 +1,7 @@
+
+
+
+class BasePage {
+
+}
+
