+++ src/po/login.page.ts
@@ -1,0 +1,5 @@
+import { BasePage } from './base.page.ts';
+
+export class LoginPage extends BasePage {
+
+}
