+++ src/po/roadmap-epam-modules/4.-Test-Automation-Framework.md
@@ -1,0 +1,33 @@
+What should be done
+
+- For this practical task, the result of the previous module should be used.
+
+Take your set up test automation framework and improve it by applying knowledge from this module. Refactor automated tests created in the previous module based on Page object pattern and follow DRY, KISS, YAGNI principles (other design patterns can be used as well)
+
+The solution should be split into next layers:
+
+1. Core layer (core functionality of TAF, that isn’t project specific).
+
+2. Business layer (should contain all functionality, related with business logic of the tested application)
+
+3. Tests layer (should contain automated tests, TAF configuration)
+
+4. Push the code to remote repository and create Merge Request
+
+
+
+Assessment criteria (pass rate is 70%)
+
+Here’s the list of actions which result in reduction of overall mark for a completed task:
+
+1. Layered architecture is not implemented (-40%)
+
+2. Page Object pattern is not used (-30%)
+
+3. Logical issues/not optimal approach (-10%)
+
+4. Code smells (naming convention, typos) (-1% per issue)
+
+5. OOP principles violations (-3% per issue)
+
+6. Code duplication is found (-3% per issue)
